Revert 35834bf2, fix tests for wfe.Challenge
